sd_ass: override converted subtitle's PlayResX to match DAR
We internally convert SRT to ASS and render them with libass, however as
of libass 0.17.0, the text shadows and borders scale in the x-axis,
while they previously did not. This means naively setting PlayRes{X,Y}
to be 4:3 will result in stretched shadows and borders in the x-axis.
Fix this by setting PlayResX = (PlayResY * DAR).